backend/requirements.txt
alembic
asgi_lifespan
argon2_cffi
asyncio
asyncpg
bcrypt
black
boto3
botocore
colorama
email-validator
fastapi
greenlet
httpx
isort
loguru
mypy
cloud-sql-python-connector[asyncpg]
passlib
pathlib
pillow
pre-commit
python-dotenv
python-decouple
python-jose
python-multipart
python-slugify
pytest
pytest-asyncio
pytest-cov
pytest-xdist
scikit-learn
SQLAlchemy
trio
uvicorn
gunicorn
password_strength
pyotp
slowapi
fastapi-mail
